pyproject.toml
[project]
name = "syntax-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "flask>=3.1.1",
    "python-dotenv>=1.1.0",
]


[tool.ruff]
line-length = 119
lint.ignore = ["F401"]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
    "ty>=0.0.1a3",
]
